Add timeout parameter and report request duration
Jenkins stopped responsing to metrics while being functional for
about 1 hour one night.
It was busy at that so I wonder if by extending the request timeout
it would have been able to get some data, and if so I want to be
able to monitor the request duration to see what it looks like.