Update rest-client to actually timeout
With version 1.8.0, the timeout was not occuring in time.

For instance with a timeout of 5 seconds, the command would just
hang and not return for quite a while.
By upgrading to 2.0.2 (latest release), this problem disappeared.

I also updated the output of metrics-jenkins.rb to print errors
in stderr and therefore keep the output graphite-compliant even
though the return is critical.