Fix issue when multiple builds are currently running
When checking the status of a build, and it is in a 'running' state,
it would check the previous build.
However there is not guarantee that this build is not running as well,
which results in a build['result'] that is nil, and an exception.

This uses the 'lastCompletedBuild' alias to get the last completed
build, since we are sure that this one is not running.